Update index.js
made changes to addNewGoal() in order to make sure no duplicate entries were made. I used array mapping in order to turn the user input goals into a list called existingGoals. I then used an if statement to check that the goalInput doesn't already exist within existingGoals and if it returns true it sends an alert and stops the duplicate from being added with return